requirements.txt
openai
fpdf2
python-dotenv
PyPDF2 